# 003-APIRESTOpenAPI/MiClienteReservas/api/openapi.yaml
openapi: 3.0.4
info:
  title: DemoApi
  version: "1.0"
servers:
- url: /
security:
- Bearer: []
...
paths:
...
  /api/products:
    get:
      operationId: GetProducts
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListApiResponse"
          description: OK
      tags:
      - Product
    post:
      operationId: CreateProduct
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/CreateProductRequestDto"
              description: Datos necesarios para crear un producto.
        required: true
      responses:
        "401":
          description: Unauthorized
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Conflict
      tags:
      - Product
  /api/products/{id}:
    delete:
      operationId: DeleteProduct
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "401":
          description: Unauthorized
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Not Found
      tags:
      - Product
    get:
      operationId: GetProductById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Not Found
      tags:
      - Product
    put:
      operationId: UpdateProduct
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/UpdateProductRequestDto"
              description: Datos que pueden modificarse de un producto existente.
        required: true
      responses:
        "401":
          description: Unauthorized
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiResponse"
          description: Conflict
      tags:
      - Product
...
components:
  schemas:
    Category:
      additionalProperties: false
      description: Representa una categoría para clasificar productos.
      example:
        id: 0
        name: name
        description: description
      properties:
        id:
          description: Identificador único de la categoría.
          format: int32
          type: integer
        name:
          description: Nombre visible de la categoría.
          nullable: true
          type: string
        description:
          description: Descripción de la categoría.
          nullable: true
          type: string
      type: object
...
    CreateProductRequestDto:
      additionalProperties: false
      description: Datos necesarios para crear un producto.
      example:
        name: name
        sku: sku
        price: 80082.82824527295
        categoryId: 1294386359
        isActive: true
      properties:
        name:
          description: Nombre del producto.
          maxLength: 200
          minLength: 2
          type: string
        sku:
          description: Código único del producto.
          maxLength: 50
          minLength: 3
          pattern: "^[A-Za-z0-9-]+$"
          type: string
        price:
          description: Precio del producto.
          format: double
          maximum: 1000000
          minimum: 0.01
          type: number
        categoryId:
          description: Identificador de la categoría del producto.
          format: int32
          maximum: 2147483647
          minimum: 1
          type: integer
        isActive:
          description: Valor que indica si el producto está activo.
          type: boolean
      required:
      - name
      - sku
      type: object
...
    Product:
      additionalProperties: false
      description: Representa un producto disponible en el catálogo.
      example:
        id: 0
        name: name
        sku: sku
        price: 6.027456183070403
        categoryId: 1
        category: ""
        isActive: true
        createdAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Identificador único del producto.
          format: int32
          type: integer
        name:
          description: Nombre visible del producto.
          nullable: true
          type: string
        sku:
          description: Código único del producto.
          nullable: true
          type: string
        price:
          description: Precio del producto.
          format: double
          type: number
        categoryId:
          description: Identificador de la categoría del producto.
          format: int32
          type: integer
        category:
          allOf:
          - $ref: "#/components/schemas/Category"
          description: Categoría a la que pertenece el producto.
          nullable: true
        isActive:
          description: Indica si el producto está activo.
          type: boolean
        createdAt:
          description: Fecha de creación del registro.
          format: date-time
          type: string
      type: object
    ProductApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data: ""
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          allOf:
          - $ref: "#/components/schemas/Product"
          description: Representa un producto disponible en el catálogo.
          nullable: true
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
    ProductListApiResponse:
      additionalProperties: false
      example:
        isSuccess: true
        message: message
        data:
        - id: 0
          name: name
          sku: sku
          price: 6.027456183070403
          categoryId: 1
          category: ""
          isActive: true
          createdAt: 2000-01-23T04:56:07.000+00:00
        - id: 0
          name: name
          sku: sku
          price: 6.027456183070403
          categoryId: 1
          category: ""
          isActive: true
          createdAt: 2000-01-23T04:56:07.000+00:00
        errors:
        - errors
        - errors
      properties:
        isSuccess:
          type: boolean
        message:
          nullable: true
          type: string
        data:
          items:
            $ref: "#/components/schemas/Product"
          nullable: true
          type: array
        errors:
          items:
            type: string
          nullable: true
          type: array
      type: object
...
    UpdateProductRequestDto:
      additionalProperties: false
      description: Datos que pueden modificarse de un producto existente.
      example:
        name: name
        sku: sku
        price: 80082.82824527295
        categoryId: 1294386359
        isActive: true
      properties:
        name:
          description: Nuevo nombre del producto.
          maxLength: 200
          minLength: 2
          type: string
        sku:
          description: Nuevo código único del producto.
          maxLength: 50
          minLength: 3
          pattern: "^[A-Za-z0-9-]+$"
          type: string
        price:
          description: Nuevo precio del producto.
          format: double
          maximum: 1000000
          minimum: 0.01
          type: number
        categoryId:
          description: Nuevo identificador de la categoría del producto.
          format: int32
          maximum: 2147483647
          minimum: 1
          type: integer
        isActive:
          description: Nuevo estado de activación del producto.
          type: boolean
      required:
      - name
      - sku
      type: object
...
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      description: "Escribe: Bearer {token}"
      scheme: bearer
      type: http